requirements.txt
grad-cam==1.4.8
huggingface-hub==0.15.1
joblib==1.2.0
jupyterlab==4.0.8
lightning==2.0.3
lit==16.0.6
matplotlib==3.7.1
notebook==6.5.4
opencv-python==4.7.0.72
openpyxl==3.1.2
pandas==2.0.2
Pillow==9.5.0
scikit-learn==1.2.2
scipy==1.10.1
timm==0.9.2
torch==2.0.1
torchaudio==2.0.2
torchmetrics==0.11.4
torchvision==0.15.2
